mobile_app/requirements.txt
flet==0.24.1
pyyaml
requests
openai
ics
